Fix arrange after cut (SPE-2583, #13637)
The cut arrange updated the Plater without the new objects
being on any of the beds. This led to mutliple beds reducing
the bed count and activating a previous bed. This in turn
ment that get_active_bed was temporarily wrong and the subsequent
arrange placed items on a wrong bed.

Improve this by explictly setting which bed is going to be
used for the arrange and than restoring the originally
active bed.

If there is just one instance, simply choose the bed the
instance is on. If there are more instances, use the active
bed for arrange.